+++ client/include/MYMQ_C.h
@@ -39,6 +39,7 @@ public:
                          ,MYMQ_Public::SupportedCallbacks cb=MYMQ_Public::CallbackNoop{});
     ClientErrorCode pull(const MYMQ_Public::TopicPartition& tp,std::vector< MYMQ_Public::ConsumerRecord>& record_batch);
 
+    ClientErrorCode seek(const MYMQ_Public::TopicPartition& tp,size_t offset_next_to_consume);
     ClientErrorCode commit_async(const MYMQ_Public::TopicPartition& tp,size_t next_offset_to_consume,MYMQ_Public::SupportedCallbacks cb=MYMQ_Public::CallbackNoop());
     void create_topic(const std::string& topicname, size_t parti_num = 1);
     void set_pull_bytes(size_t bytes);
@@ -50,7 +51,7 @@ public:
     ClientErrorCode commit_sync(const MYMQ_Public::TopicPartition& tp, size_t next_offset_to_consume);
 
     void join_group(const std::string& groupid);
-    MYMQ_Client::ClientErrorCode leave_group(const std::string& groupid);
+    ClientErrorCode leave_group(const std::string& groupid);
 
     std::unordered_set<MYMQ_Public::TopicPartition> get_assigned_partition();