+++ client/src/MYMQ_C.cpp
@@ -20,6 +20,9 @@ MYMQ_Client::ClientErrorCode MYMQ_Client::commit_async(const MYMQ_Public::TopicP
     return pimpl->commit_async(tp,next_offset_to_consume);
 }
 
+MYMQ_Client::ClientErrorCode MYMQ_Client::seek(const MYMQ_Public::TopicPartition& tp,size_t offset_next_to_consume){
+    return pimpl->seek(tp,offset_next_to_consume);
+}
 
 MYMQ_Client::ClientErrorCode MYMQ_Client::push(const MYMQ_Public::TopicPartition& tp, const std::string& key, const std::string& value
                                                ,MYMQ_Public::SupportedCallbacks cb) {